Finished Refactor
