+++ .eslintrc
@@ -1,5 +1,8 @@
 {
   "extends": [
     "wesbos"
-  ]
+  ],
+  "rules": {
+    "camelcase": "off"
+  }
 }
