+++ __tests__/userRouter.spec.js
@@ -2,7 +2,6 @@ require('dotenv').config();
 const request = require('supertest');
 const axios = require('axios');
 const server = require('../api/server');
-const db = require('../data/db-config');
 
 const badToken = 'Bearer BadToken';
 let token = null;
