+++ api/routes/routes.private.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const usersRouter = require('../../routers/userRouter');
+const usersRouter = require('../../routers/user.router');
 const { privateGrantRouter } = require('../../routers/grant.router');
 
 router.use('/users', usersRouter);
