+++ api/server.js
@@ -10,7 +10,7 @@ routes(server);
  * All 500 Errors are piped through this middleware.  This should be called 0 times with
  * proper error handling on the routes.
  */
-server.use((error, req, res) => {
+server.use((error, req, res, next) => {
   console.log('Called', error);
   res.status(500).json({ message: 'An Error Has Occurred', error });
 });
