+++ models/favorites.model.js
@@ -2,7 +2,7 @@ const db = require('../data/db-config');
 const BaseModel = require('./base.model');
 
 class FavoritesModel extends BaseModel {
-  find(id) {
+  find() {
     return db(this.name)
       .innerJoin('grants', 'grants.id', 'favorites.grant_id')
       .select('grants.*', 'favorites.id as favoriteID')
